+++ src/components/resources/index.ts
@@ -1,2 +1,1 @@
-export * from './resources-banner';
 export * from './resources-content-container';
