--- src/components/resources/resources-banner/index.ts
@@ -1,1 +1,0 @@
-export * from './resources-banner';
